Split loc in distinct files depending on the grammar
