+++ src/metrics/loc/javascript.rs
@@ -1,0 +1,83 @@
+use super::*;
+
+impl Loc for JavascriptCode {
+    fn compute(node: &Node, stats: &mut Stats, is_func_space: bool, is_unit: bool) {
+        use Javascript::*;
+
+        let (start, end) = init(node, stats, is_func_space, is_unit);
+
+        match node.object().kind_id().into() {
+            String | DQUOTE | Program => {}
+            Comment => {
+                add_cloc_lines(stats, start, end);
+            }
+            ExpressionStatement | ExportStatement | ImportStatement | StatementBlock
+            | IfStatement | SwitchStatement | ForStatement | ForInStatement | WhileStatement
+            | DoStatement | TryStatement | WithStatement | BreakStatement | ContinueStatement
+            | DebuggerStatement | ReturnStatement | ThrowStatement | EmptyStatement
+            | StatementIdentifier => {
+                stats.logical_lines += 1;
+            }
+            _ => {
+                check_comment_ends_on_code_line(stats, start);
+                stats.lines.insert(start);
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::path::PathBuf;
+
+    use super::*;
+
+    #[test]
+    fn javascript_no_zero_blank() {
+        // Checks that the blank metric is not equal to 0 when there are some
+        // comments next to code lines.
+        check_metrics!(
+            "function ConnectToUpdateServer() {
+              var pool = 0;
+
+              var updateServer = -42;
+              var isConnected = false;
+              var currTry = 0;
+              var numRetries = 10;  // Number of IPC connection retries before
+                                    // giving up.
+              var numTries = 20;    // Number of IPC connection tries before
+                                    // giving up.
+            }",
+            "foo.js",
+            JavascriptParser,
+            loc,
+            [
+                (sloc, 11, usize), // The number of lines is 11
+                (ploc, 8, usize),  // The number of code lines is 8
+                (cloc, 4, usize),  // The number of comments is 4
+                (blank, 1, usize)  // The number of blank lines is 1
+            ]
+        );
+    }
+
+    #[test]
+    fn javascript_real_loc() {
+        check_metrics!(
+            "assert.throws(Test262Error, function() {
+               for (let { poisoned: x = ++initEvalCount } = poisonedProperty; ; ) {
+                 return;
+               }
+             });",
+            "foo.js",
+            JavascriptParser,
+            loc,
+            [
+                (sloc, 5, usize),  // The number of lines is 5
+                (ploc, 5, usize),  // The number of code lines is 5
+                (lloc, 7, usize),  // The number of statements is 7
+                (cloc, 0, usize),  // The number of comments is 0
+                (blank, 0, usize)  // The number of blank lines is 0
+            ]
+        );
+    }
+}
