+++ src/metrics/loc/mod.rs
@@ -1,0 +1,200 @@
+use std::fmt;
+
+use fxhash::FxHashSet;
+use serde::ser::{SerializeStruct, Serializer};
+use serde::Serialize;
+
+use crate::checker::Checker;
+use crate::*;
+
+mod cpp;
+mod javascript;
+mod mozjs;
+mod python;
+mod rust;
+mod tsx;
+mod typescript;
+
+/// The `Loc` metric suite.
+#[derive(Debug, Clone, Default)]
+pub struct Stats {
+    start: usize,
+    end: usize,
+    unit: bool,
+    lines: FxHashSet<usize>,
+    logical_lines: usize,
+    only_comment_lines: usize,
+    code_comment_lines: usize,
+    comment_line_end: Option<usize>,
+}
+
+impl Serialize for Stats {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut st = serializer.serialize_struct("loc", 5)?;
+        st.serialize_field("sloc", &self.sloc())?;
+        st.serialize_field("ploc", &self.ploc())?;
+        st.serialize_field("lloc", &self.lloc())?;
+        st.serialize_field("cloc", &self.cloc())?;
+        st.serialize_field("blank", &self.blank())?;
+        st.end()
+    }
+}
+
+impl fmt::Display for Stats {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "sloc: {}, ploc: {}, lloc: {}, cloc: {}, blank: {}",
+            self.sloc(),
+            self.ploc(),
+            self.lloc(),
+            self.cloc(),
+            self.blank(),
+        )
+    }
+}
+
+impl Stats {
+    /// Merges a second `Loc` metric suite into the first one
+    pub fn merge(&mut self, other: &Stats) {
+        // Merge ploc lines
+        for l in other.lines.iter() {
+            self.lines.insert(*l);
+        }
+
+        // Merge lloc lines
+        self.logical_lines += other.logical_lines;
+
+        // Merge cloc lines
+        self.only_comment_lines += other.only_comment_lines;
+        self.code_comment_lines += other.code_comment_lines;
+    }
+
+    /// The `Sloc` metric.
+    ///
+    /// Counts the number of lines in a scope
+    #[inline(always)]
+    pub fn sloc(&self) -> f64 {
+        // This metric counts the number of lines in a file
+        // The if construct is needed to count the line of code that represents
+        // the function signature in a function space
+        let sloc = if self.unit {
+            self.end - self.start
+        } else {
+            (self.end - self.start) + 1
+        };
+        sloc as f64
+    }
+
+    /// The `Ploc` metric.
+    ///
+    /// Counts the number of instruction lines in a scope
+    #[inline(always)]
+    pub fn ploc(&self) -> f64 {
+        // This metric counts the number of instruction lines in a code
+        // https://en.wikipedia.org/wiki/Source_lines_of_code
+        self.lines.len() as f64
+    }
+
+    /// The `Lloc` metric.
+    ///
+    /// Counts the number of statements in a scope
+    #[inline(always)]
+    pub fn lloc(&self) -> f64 {
+        // This metric counts the number of statements in a code
+        // https://en.wikipedia.org/wiki/Source_lines_of_code
+        self.logical_lines as f64
+    }
+
+    /// The `Cloc` metric.
+    ///
+    /// Counts the number of comments in a scope
+    #[inline(always)]
+    pub fn cloc(&self) -> f64 {
+        // Comments are counted regardless of their placement
+        // https://en.wikipedia.org/wiki/Source_lines_of_code
+        (self.only_comment_lines + self.code_comment_lines) as f64
+    }
+
+    /// The `Blank` metric.
+    ///
+    /// Counts the number of blank lines in a scope
+    #[inline(always)]
+    pub fn blank(&self) -> f64 {
+        self.sloc() - self.ploc() - self.only_comment_lines as f64
+    }
+}
+
+#[doc(hidden)]
+pub trait Loc
+where
+    Self: Checker,
+{
+    fn compute(_node: &Node, _stats: &mut Stats, _is_func_space: bool, _is_unit: bool) {}
+}
+
+#[inline(always)]
+pub(crate) fn init(
+    node: &Node,
+    stats: &mut Stats,
+    is_func_space: bool,
+    is_unit: bool,
+) -> (usize, usize) {
+    let start = node.object().start_position().row;
+    let end = node.object().end_position().row;
+
+    if is_func_space {
+        stats.start = start;
+        stats.end = end;
+        stats.unit = is_unit;
+    }
+    (start, end)
+}
+
+#[inline(always)]
+// Discriminates among the comments that are *after* a code line and
+// the ones that are on an independent line.
+// This difference is necessary in order to avoid having
+// a wrong count for the blank metric.
+pub(crate) fn add_cloc_lines(stats: &mut Stats, start: usize, end: usize) {
+    let comment_diff = end - start;
+    let is_comment_after_code_line = stats.lines.contains(&start);
+    if is_comment_after_code_line && comment_diff == 0 {
+        // A comment is *entirely* next to a code line
+        stats.code_comment_lines += 1;
+    } else if is_comment_after_code_line && comment_diff > 0 {
+        // A block comment that starts next to a code line and ends on
+        // independent lines.
+        stats.code_comment_lines += 1;
+        stats.only_comment_lines += comment_diff;
+    } else {
+        // A comment on an independent line AND
+        // a block comment on independent lines OR
+        // a comment *before* a code line
+        stats.only_comment_lines += (end - start) + 1;
+        // Save line end of a comment to check whether
+        // a comment *before* a code line is considered
+        stats.comment_line_end = Some(end);
+    }
+}
+
+#[inline(always)]
+// Detects the comments that are on a code line but *before* the code part.
+// This difference is necessary in order to avoid having
+// a wrong count for the blank metric.
+pub(crate) fn check_comment_ends_on_code_line(stats: &mut Stats, start_code_line: usize) {
+    if let Some(end) = stats.comment_line_end {
+        if end == start_code_line && !stats.lines.contains(&start_code_line) {
+            // Comment entirely *before* a code line
+            stats.only_comment_lines -= 1;
+            stats.code_comment_lines += 1;
+        }
+    }
+}
+
+impl Loc for PreprocCode {}
+impl Loc for CcommentCode {}
+impl Loc for JavaCode {}
