+++ src/metrics/loc/mozjs.rs
@@ -1,0 +1,55 @@
+use super::*;
+
+impl Loc for MozjsCode {
+    fn compute(node: &Node, stats: &mut Stats, is_func_space: bool, is_unit: bool) {
+        use Mozjs::*;
+
+        let (start, end) = init(node, stats, is_func_space, is_unit);
+
+        match node.object().kind_id().into() {
+            String | DQUOTE | Program => {}
+            Comment => {
+                add_cloc_lines(stats, start, end);
+            }
+            ExpressionStatement | ExportStatement | ImportStatement | StatementBlock
+            | IfStatement | SwitchStatement | ForStatement | ForInStatement | WhileStatement
+            | DoStatement | TryStatement | WithStatement | BreakStatement | ContinueStatement
+            | DebuggerStatement | ReturnStatement | ThrowStatement | EmptyStatement
+            | StatementIdentifier => {
+                stats.logical_lines += 1;
+            }
+            _ => {
+                check_comment_ends_on_code_line(stats, start);
+                stats.lines.insert(start);
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::path::PathBuf;
+
+    use super::*;
+
+    #[test]
+    fn mozjs_real_loc() {
+        check_metrics!(
+            "assert.throws(Test262Error, function() {
+               for (let { poisoned: x = ++initEvalCount } = poisonedProperty; ; ) {
+                 return;
+               }
+             });",
+            "foo.js",
+            MozjsParser,
+            loc,
+            [
+                (sloc, 5, usize),  // The number of lines is 5
+                (ploc, 5, usize),  // The number of code lines is 5
+                (lloc, 7, usize),  // The number of statements is 7
+                (cloc, 0, usize),  // The number of comments is 0
+                (blank, 0, usize)  // The number of blank lines is 0
+            ]
+        );
+    }
+}
