+++ src/metrics/loc/tsx.rs
@@ -1,0 +1,27 @@
+use super::*;
+
+impl Loc for TsxCode {
+    fn compute(node: &Node, stats: &mut Stats, is_func_space: bool, is_unit: bool) {
+        use Tsx::*;
+
+        let (start, end) = init(node, stats, is_func_space, is_unit);
+
+        match node.object().kind_id().into() {
+            String | DQUOTE | Program => {}
+            Comment => {
+                add_cloc_lines(stats, start, end);
+            }
+            ExpressionStatement | ExportStatement | ImportStatement | StatementBlock
+            | IfStatement | SwitchStatement | ForStatement | ForInStatement | WhileStatement
+            | DoStatement | TryStatement | WithStatement | BreakStatement | ContinueStatement
+            | DebuggerStatement | ReturnStatement | ThrowStatement | EmptyStatement
+            | StatementIdentifier => {
+                stats.logical_lines += 1;
+            }
+            _ => {
+                check_comment_ends_on_code_line(stats, start);
+                stats.lines.insert(start);
+            }
+        }
+    }
+}
